+++ jvm-client/src/main/kotlin/com/tomaszezula/makker/client/jvm/java/model/ModuleUpdate.kt
@@ -1,3 +1,3 @@
 package com.tomaszezula.makker.client.jvm.java.model
 
-data class ModuleUpdate(val moduleId: Long, val key: String, val value: String)
+data class ModuleUpdate(val moduleId: Int, val key: String, val value: String)
